DIRECTOR: LINGO: Move char normalizations to lingo-utils.cpp
